Changed to long so they can be beyond 2G

+++ WebServerStartupInfo.cs
@@ -59,7 +59,7 @@ namespace Api.Startup
 			services.Configure<FormOptions>(x =>
 			{
 				x.ValueLengthLimit = int.MaxValue;
-				x.MultipartBodyLengthLimit = int.MaxValue; // In case of multipart
+				x.MultipartBodyLengthLimit = long.MaxValue; // In case of multipart
 			});
 			
 			// Start checking types:
